# backend/backend-java/src/main/java/com/backend/model/Message.java
package com.backend.model;

import jakarta.persistence.*;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;
import org.hibernate.annotations.CreationTimestamp;

import java.time.LocalDateTime;

@Entity
@Table(name = "MESSAGES")
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Message {
    
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "MESSAGEID")
    private Long messageId;
    
    @NotBlank(message = "Message content is required")
    @Size(max = 2000, message = "Message content must not exceed 2000 characters")
    @Column(name = "CONTENT", length = 2000)
    private String content;
    
    @CreationTimestamp
    @Column(name = "TIMESTAMP")
    private LocalDateTime timestamp;
    
    @Column(name = "ISREAD")
    private Boolean isRead = false;
    
    // Relationships
    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "SENDERID", referencedColumnName = "USERID")
    @NotNull
    private User sender;
    
    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "RECEIVERID", referencedColumnName = "USERID")
    @NotNull
    private User receiver;
} 